* 'master' of github.com:scylladb/scylla:
  Additional entries in CODEOWNERS
  tests: add large paging state tests
  cdc: add an abstraction for building log mutations
  Update seastar submodule
  Simplify a few calls to seastar::make_shared
  Use detect_stack_use_after_return=1
  alternator::streams: Don't include empty new/old image
  alternator::streams: Always include "Records" array in get_records reponse
  alternator::streams: Incr shard iterator threshold in get_records
  increase the maximum size of query results to 2^64
  sstables: Delete write_failed
  sstables: Move monitor after writer in compaction_writer

+++ querier.hh
@@ -68,7 +68,7 @@ auto consume_page(flat_mutation_reader& reader,
         lw_shared_ptr<compact_for_query_state<OnlyLive>> compaction_state,
         const query::partition_slice& slice,
         Consumer&& consumer,
-        uint32_t row_limit,
+        uint64_t row_limit,
         uint32_t partition_limit,
         gc_clock::time_point query_time,
         db::timeout_clock::time_point timeout,
@@ -208,7 +208,7 @@ public:
     template <typename Consumer>
     requires CompactedFragmentsConsumer<Consumer>
     auto consume_page(Consumer&& consumer,
-            uint32_t row_limit,
+            uint64_t row_limit,
             uint32_t partition_limit,
             gc_clock::time_point query_time,
             db::timeout_clock::time_point timeout,
